Applying the ZERO-specific setup [ZER-516]

+++ src/fountain-utils.cpp
@@ -15,7 +15,7 @@ namespace ur {
 
 size_t choose_degree(size_t seq_len, Xoshiro256& rng) {
     vector<double> degree_probabilities;
-    for(int i = 1; i <= seq_len; i++) {
+    for(size_t i = 1; i <= seq_len; i++) {
         degree_probabilities.push_back(1.0 / i);
     }
     auto degree_chooser = RandomSampler(degree_probabilities);
@@ -34,7 +34,7 @@ set<size_t> choose_fragments(uint32_t seq_num, size_t seq_len, uint32_t checksum
         auto degree = choose_degree(seq_len, rng);
         vector<size_t> indexes;
         indexes.reserve(seq_len);
-        for(int i = 0; i < seq_len; i++) { indexes.push_back(i); }
+        for(size_t i = 0; i < seq_len; i++) { indexes.push_back(i); }
         auto shuffled_indexes = shuffled(indexes, rng);
         return set<size_t>(shuffled_indexes.begin(), shuffled_indexes.begin() + degree);
     }
